+++ backend/src/app.test.ts
@@ -1,8 +1,8 @@
 describe('最初のテスト', () => {
   it('最初のテストとその実装', () => {
     const statusCode = 201
-    expect(statusCode).toBe(201)
+    // expect(statusCode).toBe(201)
 
-    // expect(20 + 50).not.toBe(70)
+    expect(20 + 50).toBe(70)
   })
 })
